+++ mepo.d/command/compare/compare.py
@@ -8,8 +8,24 @@ VER_LEN = 30
 
 def run(args):
     allcomps = MepoState.read_state()
-    max_namelen, max_origlen = calculate_header_lengths(allcomps)
-    print_header(max_namelen, max_origlen)
+
+    if not any_differing_repos(allcomps):
+        print(f'No repositories have changed')
+    else:
+        max_namelen, max_origlen = calculate_header_lengths(allcomps)
+        print_header(max_namelen, max_origlen)
+        for comp in allcomps:
+            git = GitRepository(comp.remote, comp.local)
+            curr_ver = version_to_string(git.get_version(),git)
+            orig_ver = version_to_string(comp.version,git)
+
+            # This command is to try and work with git tag oddities
+            curr_ver = sanitize_version_string(orig_ver,curr_ver,git)
+
+            print_cmp(comp.name, orig_ver, curr_ver, max_namelen, max_origlen, args.all)
+
+def any_differing_repos(allcomps):
+
     for comp in allcomps:
         git = GitRepository(comp.remote, comp.local)
         curr_ver = version_to_string(git.get_version(),git)
@@ -18,7 +34,10 @@ def run(args):
         # This command is to try and work with git tag oddities
         curr_ver = sanitize_version_string(orig_ver,curr_ver,git)
 
-        print_cmp(comp.name, orig_ver, curr_ver, max_namelen, max_origlen, args.brief)
+        if curr_ver not in orig_ver:
+            return True
+
+    return False
 
 def calculate_header_lengths(allcomps):
     names = []
@@ -37,7 +56,7 @@ def print_header(max_namelen, max_origlen):
     print(FMTHEAD.format("Repo","Original","Current"))
     print(FMTHEAD.format("-"*80,"-"*max_origlen,"-"*7))
 
-def print_cmp(name, orig, curr, name_width, orig_width, brief):
+def print_cmp(name, orig, curr, name_width, orig_width, all_repos):
     name_blank = ''
     #if orig not in curr:
     if curr not in orig:
@@ -45,8 +64,8 @@ def print_cmp(name, orig, curr, name_width, orig_width, brief):
         name_blank = colors.RED + name_blank + colors.RESET
         name_width += len(colors.RED) + len(colors.RESET)
     else:
-        # This only prints differing repos if --brief is passed in
-        if brief:
+        # This only prints differing repos unless --all is passed in
+        if not all_repos:
             return
     FMT_VAL = (name_width, name_width, orig_width)
 
@@ -61,3 +80,4 @@ def print_cmp(name, orig, curr, name_width, orig_width, brief):
         print(FMT2.format(name_blank, '...', curr))
     else:
         print(FMT0.format(name, orig, curr))
+